requirements.txt
tkinter-tooltip==2.2.0
Pillow==10.1.0
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
supabase==2.3.0
python-dotenv==1.0.0
